+++ src/clients/AcrossConfigStoreClient/AcrossConfigStoreClient.ts
@@ -15,7 +15,6 @@ import {
   max,
   sortEventsAscending,
   findLast,
-  isUBA,
 } from "../../utils";
 import { Contract, BigNumber, Event } from "ethers";
 import winston from "winston";
@@ -234,32 +233,6 @@ export class AcrossConfigStoreClient extends BaseAbstractClient {
     const version = this.getConfigStoreVersionForTimestamp(timestamp);
     return this.isValidConfigStoreVersion(version);
   }
-
-  /**
-   * Return first block number where UBA was activated by setting "Version" GlobalConfig in ConfigStore
-   * @returns
-   */
-  getUbaActivationBlock(): number {
-    const config = this.cumulativeConfigStoreVersionUpdates.find((config) => {
-      isUBA(Number(config.value));
-    });
-    if (config) {
-      return config.blockNumber;
-    } else {
-      return Number.MAX_SAFE_INTEGER;
-    }
-  }
-
-  public isUbaBlock(block: number): boolean {
-    const versionAppliedToDeposit = this.getConfigStoreVersionForBlock(block);
-    return isUBA(versionAppliedToDeposit);
-  }
-
-  public canSupportUbaBlock(block: number): boolean {
-    const versionAppliedToDeposit = this.getConfigStoreVersionForBlock(block);
-    return this.isValidConfigStoreVersion(versionAppliedToDeposit);
-  }
-
   isValidConfigStoreVersion(version: number): boolean {
     return this.configStoreVersion >= version;
   }
